+++ changelog.rst
@@ -29,6 +29,7 @@ Important updates:
 - SPINNA: GUI single sim does not allow the sum of proportions to exceed 100% (see issue #560)
 - SPINNA: save last opened folder added
 - SPINNA: save NND plot fixed (when no simulations were run)
+- Simplify the API for picking similar in ``picasso.postprocess``
 
 0.8.8
 -----